+++ src/app/Http/Controllers/HelplineController.php
@@ -246,7 +246,9 @@ class HelplineController extends Controller
     {
         if ($request->has('noop')) {
             if ($request->has('CallStatus') &&
-                ($request->get('CallStatus') == TwilioCallStatus::NOANSWER || $request->get('CallStatus') == TwilioCallStatus::FAILED)) {
+                ($request->get('CallStatus') == TwilioCallStatus::NOANSWER
+                    || $request->get('CallStatus') == TwilioCallStatus::FAILED
+                    || $request->get('CallStatus') == TwilioCallStatus::BUSY)) {
                 $this->twilio->incrementNoAnswerCount();
             }
 
@@ -301,7 +303,11 @@ class HelplineController extends Controller
 
             // TODO: Make timeout configurable per volunteer
             if (( $request->has('SequenceNumber') && intval($request->get('SequenceNumber')) == 1 ) ||
-                ( $request->has('CallStatus') && ($request->get('CallStatus') == TwilioCallStatus::NOANSWER || $request->get('CallStatus') == TwilioCallStatus::COMPLETED || $request->get('CallStatus') == TwilioCallStatus::FAILED ))) {
+                ( $request->has('CallStatus') &&
+                    ($request->get('CallStatus') == TwilioCallStatus::NOANSWER
+                        || $request->get('CallStatus') == TwilioCallStatus::COMPLETED
+                        || $request->get('CallStatus') == TwilioCallStatus::FAILED
+                        || $request->get('CallStatus') == TwilioCallStatus::BUSY ))) {
                 $callConfig = $this->getCallConfig($request, $serviceBodyCallHandling);
 
                 if ($request->has('CallStatus') && $request->get('CallStatus') == TwilioCallStatus::NOANSWER) {
@@ -314,6 +320,14 @@ class HelplineController extends Controller
                 } else if ($request->has('CallStatus') && $request->get('CallStatus') == TwilioCallStatus::FAILED) {
                     Log::error(sprintf("Volunteer Call Failed %s: %s", $request->get('Called'), $request->get('ErrorMessage')));
                     $this->call->insertCallEventRecord(EventId::VOLUNTEER_NUMBER_BAD, (object)['to_number' => $request->get('Called'), 'error' => $request->get('ErrorMessage')]);
+                    $this->call->setConferenceParticipant(
+                        $request->get('FriendlyName'),
+                        $request->get('CallSid'),
+                        CallRole::VOLUNTEER
+                    );
+                } else if ($request->has('CallStatus') && $request->get('CallStatus') == TwilioCallStatus::BUSY) {
+                    Log::error(sprintf("Volunteer Call Busy %s: %s", $request->get('Called'), $request->get('ErrorMessage')));
+                    $this->call->insertCallEventRecord(EventId::VOLUNTEER_NUMBER_BUSY, (object)['to_number' => $request->get('Called')]);
                     $this->call->setConferenceParticipant(
                         $request->get('FriendlyName'),
                         $request->get('CallSid'),