+++ src/app/Models/ConfigData.php
@@ -54,12 +54,25 @@ class ConfigData extends Model
     public static function createVolunteer(
         int $serviceBodyId,
         int $parentServiceBodyId,
-        object $volunteerConfiguration
+        VolunteerData $volunteerConfiguration
     ) : void {
         self::create([
             "service_body_id"=>$serviceBodyId,
             "parent_id"=>$parentServiceBodyId,
             "data"=>json_encode([$volunteerConfiguration]),
+            "data_type"=>DataType::YAP_VOLUNTEERS_V2
+        ]);
+    }
+
+    public static function createVolunteers(
+        int $serviceBodyId,
+        int $parentServiceBodyId,
+        array $volunteerDataArray
+    ) : void {
+        self::create([
+            "service_body_id"=>$serviceBodyId,
+            "parent_id"=>$parentServiceBodyId,
+            "data"=>json_encode($volunteerDataArray),
             "data_type"=>DataType::YAP_VOLUNTEERS_V2
         ]);
     }